User Authentication and signup and signinTested

+++ src/main/java/com/example/bookrent/Service/ReviewService.java
@@ -5,7 +5,7 @@ import com.example.bookrent.Dto.ReviewDto;
 import java.util.List;
 
 public interface ReviewService {
-    public ReviewDto submitReview(Long userId, Long bookId, Integer rating, String reviewText);
+    public ReviewDto submitReview(Long userId, Long bookId, Double rating, String reviewText);
     public List<ReviewDto> getReviewsForBook(Long bookId);
     public Double getAverageRatingForBook(Long bookId);
 }
